Backport coroutines having their own file and luaopen_* function
Keeping coroutines in the base library file makes diffs with newer Lua version unnecessarily messy. Coroutines for now haven't been changed much, the Lua 5.2 logic is subtly different and there's no pressing need to backport it.
