+++ lib/lua/src/lualib.h
@@ -15,8 +15,10 @@
 #define LUA_FILEHANDLE		"FILE*"
 
 
+LUALIB_API int (luaopen_base)(lua_State* L);
+
 #define LUA_COLIBNAME	"coroutine"
-LUALIB_API int (luaopen_base) (lua_State *L);
+LUALIB_API int (luaopen_coroutine)(lua_State* L);
 
 #define LUA_TABLIBNAME	"table"
 LUALIB_API int (luaopen_table) (lua_State *L);
